Add OIDC client CRD + restructure Keycloak client
